🔨 refactor: my place 코드 분리
