+++ src/app/(user)/(service)/my-place/layout.tsx
@@ -1,5 +1,11 @@
 import { ReactNode } from "react";
 
 export default function Layout({ children }: { children: ReactNode }) {
-  return <div>{children}</div>;
+  return (
+    <div className="flex justify-center">
+      <div className="flex w-full max-w-[1240px] flex-col items-center justify-center">
+        {children}
+      </div>
+    </div>
+  );
 }
