+++ src/features/my-place/my-place-edit/actions/use-delete-bookmarks.ts
@@ -1,0 +1,18 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import client from "@/src/lib/api/client";
+
+export const useDeleteBookmarks = (memberId?: number) => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: async (bookmarkIds: number[]) => {
+      if (!memberId) return;
+      await client.delete(`/api/members/${memberId}/bookmarks`, {
+        params: { bookmarkIds: bookmarkIds.join(",") },
+      });
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ["myPlaces", memberId] });
+    },
+  });
+};
