+++ src/features/my-place/my-place-edit/actions/use-reorder-bookmarks.ts
@@ -1,0 +1,20 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import client from "@/src/lib/api/client";
+import { ReorderBody } from "@/src/types/my-place";
+
+export const useReorderBookmarks = (memberId?: number) => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: async (body: ReorderBody[]) => {
+      if (!memberId) return;
+      return await client.patch(
+        `/api/members/${memberId}/bookmarks/sequence/reorder`,
+        body,
+      );
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ["myPlaces", memberId] });
+    },
+  });
+};
