+++ src/features/my-place/my-place-view/components/my-place-item.tsx
